refactor(demo): extract examples + new blankslate example
